move mesh conversion functions to mesh_utils.py
